feeder/requirements.txt
paho-mqtt==1.5.0
chromedriver-autoinstaller==0.6.4
selenium==4.23.1
requests==2.32.3
matplotlib==3.9.2
cartopy==0.23.0
numpy==2.0.1
redis==5.0.8
async-timeout==4.0.3
Pillow
flask==3.0.3